requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
certifi==2022.9.24
charset-normalizer==2.0.9
dateparser==1.1.0
frozenlist==1.4.1
greenlet==1.1.3.post0
idna==3.3
multidict==5.2.0
psycopg2-binary==2.9.9
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.42
typing_extensions==4.3.0
tzdata==2021.5
tzlocal==4.1
ujson==4.3.0
urllib3==1.26.7
websocket-client==1.2.1
websockets==9.1
yarl==1.9.4
